backend/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
#mysqlclient==2.1.1
paramiko==2.12.0
python-crontab==2.6.0
django-filter==23.1